pyproject.toml
[tool.poetry]
name = "cafe-app"
version = "0.1.0"
description = ""
authors = ["rukiqq <mitaka@duck.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = "~4.2.19"
djangorestframework = "^3.15.2"
django-filter = "^25.1"


[tool.poetry.group.dev.dependencies]
black = "^25.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
